// src/types/TablePageTableInfo.ts
type TablePageTableInfo = {
  id: string;
  name: string;
  visibility?: string;
  description?: string;
  type?: string;
  created?: string;
  updated?: string;  
  lines?: string;
  owner?: string;
  table_name?: string;
  org_id?: string;
  app_name?: string;
  created_at?: string;
  updated_at?: string;
  linecount?: string;
};;


export default TablePageTableInfo;