// src/types/TablePageEditTableType.ts
type TablePageEditTableType = {
    visibility: string;
    columnNames: string;
    columns: []
    delimiter?: string;
    downloadUrl?: string;
    accountID?: string;
    columnCount?: string;
    storeTableDataAsFile?: string;
    appId?: string;
    delimiterDescription?: string;
    created?: string;
    updated?: string;
    name?: string;
    id?: string;
    type?: string;
    size?: string;
    owner?: string;
    shared?: string;
    description?: string;
    editable: string;
    isOwner: string;
    rowSummary: [[]];

};

export default TablePageEditTableType;